feat: switch to always return a string
Previously, using png extension results in a return value of type buffer for the user to deal with.
However, this doesn't make sense since this package already assumes that the user eventually wants
an html element to return. Therefore, png images are converted into base64 source url and inserted
into an image tag. The output will be a string regardless of the extension. The output is always a
valid HTML

BREAKING CHANGE: The function will no longer return a Buffer, it will instead return a string
containing an image tag with the image inserted as base64 source url